Manage errors, header title, and filling in library questions in editor manager
